+++ VsCSharpWinForm_sample2/VsCSharpWinForm_sample2/FrmMain.cs
@@ -230,20 +230,20 @@ namespace VsCSharpWinForm_sample2
                     char[] cArrayTrim = { ' ', '\t', (char)10, (char)13 };
 
                     /// system.
-                    Param.IsService = TIniFile.GetBool(TIniFile.DoIniParaInBuffer(false, sIni, "IsService", "", Param.IsService.ToString()).Trim(cArrayTrim).ToUpper());
+                    Param.IsService = FileHelper.INI.GetBool(FileHelper.INI.DoIniParaInBuffer(false, sIni, "IsService", "", Param.IsService.ToString()).Trim(cArrayTrim).ToUpper());
                     //GeneralT.DoIniParaInBufferGetInteger(ref miMainLoopSleepingIntervalInMS, sIni, "MainLoopSleepingIntervalInMS", "", miMainLoopSleepingIntervalInMS);
 
                     /// Log.
                     if (Logger != null)
                     {
-                        Logger.FilePathFormat = GeneralT.GetDefaultAbsolutePathIfRelative(TIniFile.DoIniParaInBuffer(false, sIni, "FilePathFormat", "", Logger.FilePathFormat).Trim(cArrayTrim));
-                        Logger.ContentFormat = TIniFile.DoIniParaInBuffer(false, sIni, "ContentFormat", "", Logger.ContentFormat).Trim(cArrayTrim);
+                        Logger.FilePathFormat = GeneralT.GetDefaultAbsolutePathIfRelative(FileHelper.INI.DoIniParaInBuffer(false, sIni, "FilePathFormat", "", Logger.FilePathFormat).Trim(cArrayTrim));
+                        Logger.ContentFormat = FileHelper.INI.DoIniParaInBuffer(false, sIni, "ContentFormat", "", Logger.ContentFormat).Trim(cArrayTrim);
                     }
 
                     /// Login.
-                    Param.Login.MaxRetry = TIniFile.GetInt(TIniFile.DoIniParaInBuffer(false, sIni, "LoginMaxRetry", "", Param.Login.MaxRetry.ToString()), 0) ?? 0;
-                    Param.Login.FailMessage = TIniFile.DoIniParaInBuffer(false, sIni, "LoginFailMessage", "", Param.Login.FailMessage).Trim(cArrayTrim);
-                    Param.Login.ExceedMaxRetryMessage = TIniFile.DoIniParaInBuffer(false, sIni, "LoginExceedMaxRetryMessage", "", Param.Login.ExceedMaxRetryMessage).Trim(cArrayTrim);
+                    Param.Login.MaxRetry = FileHelper.INI.GetInt(FileHelper.INI.DoIniParaInBuffer(false, sIni, "LoginMaxRetry", "", Param.Login.MaxRetry.ToString()), 0) ?? 0;
+                    Param.Login.FailMessage = FileHelper.INI.DoIniParaInBuffer(false, sIni, "LoginFailMessage", "", Param.Login.FailMessage).Trim(cArrayTrim);
+                    Param.Login.ExceedMaxRetryMessage = FileHelper.INI.DoIniParaInBuffer(false, sIni, "LoginExceedMaxRetryMessage", "", Param.Login.ExceedMaxRetryMessage).Trim(cArrayTrim);
                 }
             }
             catch (Exception ex) { Logger?.Error(ex); }
@@ -290,11 +290,12 @@ namespace VsCSharpWinForm_sample2
                 string exeFolder = System.IO.Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location);
                 System.IO.Directory.SetCurrentDirectory(exeFolder);
 
-                TIniFile.Logger = Logger;
+                //FileHelper.INI.Logger = Logger;
                 /// Declare instances of the objects.
 
                 ReadIni();/// Read initial file.
                 /// Set the logger of objects.
+                FileHelper.Logger = Logger;
                 CyptoRijndaelT.Logger = Logger;
                 GeneralT.Logger = Logger;
                 TTcpClientSocket.Logger = Logger;
@@ -305,9 +306,6 @@ namespace VsCSharpWinForm_sample2
                 Views.FrmWait.Logger = Logger;
                 Views.FrmTcpClient.Logger = Logger;
                 MailHelper.Logger = Logger;
-                ZipHelper.Logger = Logger;
-                FileTransferHelper.Logger = Logger;
-                TCsvFile.Logger = Logger;
                 ExcelHelper.Logger = Logger;
 
                 /// Show the help message if run in the command line with proper switch.
@@ -447,7 +445,9 @@ namespace VsCSharpWinForm_sample2
             try
             {
                 WriteLogToUI("1234984 64633 d2g");
-                WriteLogToUI("abcdef {0:HH:mm:ss.fff}", DateTime.Now);
+                DateTime t = DateTime.Now;
+                WriteLogToUI("Now = {0:yyyy-MM-dd HH:mm:ss.fffffff}", t);
+                WriteLogToUI("UTC = {0:yyyy-MM-ddTHH:mm:ss.fffffffZ}", t.ToUniversalTime());
             }
             catch (Exception ex) { Logger?.Error(ex); }
         }
@@ -473,7 +473,7 @@ namespace VsCSharpWinForm_sample2
         {
             try
             {
-                string[] files = GeneralT.GetFilesWithPatternAndExpiryDayCountInFolder("log", null, null, ".log", 2);
+                string[] files = GeneralT.GetFilesWithPatternAndExpiryDayCountInFolder("log", null, null, ".log", 2, "yyyy-MM-dd");
                 if ((files?.Length ?? 0) > 0)
                 {
                     LocalLogger(TLog.LogLevel.INFO, "FileCount = {0}", files.Length);
@@ -1500,7 +1500,7 @@ namespace VsCSharpWinForm_sample2
                     GeneralT.GetDefaultAbsolutePathIfRelative("x64")
                 };
                 LocalLogger(TLog.LogLevel.DEBUG, "Source Paths:{0}{1}", Environment.NewLine, string.Join(Environment.NewLine, sourcePaths));
-                if (ZipHelper.DotNet.Zip(zipFilepath, password, sourcePaths))
+                if (FileHelper.Zip.DotNet.Zip(zipFilepath, password, sourcePaths))
                     LocalLogger(TLog.LogLevel.DEBUG, "Succeed to zip items.");
                 else
                 {
@@ -1508,7 +1508,7 @@ namespace VsCSharpWinForm_sample2
                     return;
                 }
                 string extractedFolder = GeneralT.GetDefaultAbsolutePathIfRelative("extract");
-                if (ZipHelper.DotNet.Unzip(zipFilepath, password, extractedFolder))
+                if (FileHelper.Zip.DotNet.Unzip(zipFilepath, password, extractedFolder))
                     LocalLogger(TLog.LogLevel.DEBUG, "Succeed to unzip items.");
                 else
                 {
@@ -1518,7 +1518,7 @@ namespace VsCSharpWinForm_sample2
                 LocalLogger(TLog.LogLevel.DEBUG, "7 zip");
                 LocalLogger(TLog.LogLevel.DEBUG, "Source Paths:{0}{1}", Environment.NewLine, string.Join(Environment.NewLine, sourcePaths));
                 zipFilepath = GeneralT.GetDefaultAbsolutePathIfRelative("a2.zip");
-                if (ZipHelper.SevenZip.Zip(zipFilepath, password, out string output, out string error, sourcePaths))
+                if (FileHelper.Zip.SevenZip.Zip(zipFilepath, password, out string output, out string error, sourcePaths))
                     LocalLogger(TLog.LogLevel.DEBUG, "Succeed to zip items.");
                 else
                 {
@@ -1526,7 +1526,7 @@ namespace VsCSharpWinForm_sample2
                     return;
                 }
                 extractedFolder = GeneralT.GetDefaultAbsolutePathIfRelative("extract2");
-                if (ZipHelper.SevenZip.Unzip(zipFilepath, password, extractedFolder, out output, out error))
+                if (FileHelper.Zip.SevenZip.Unzip(zipFilepath, password, extractedFolder, out output, out error))
                     LocalLogger(TLog.LogLevel.DEBUG, "Succeed to unzip items.");
                 else
                 {
@@ -1658,7 +1658,7 @@ namespace VsCSharpWinForm_sample2
                 //    }
                 //}
 
-                //List<List<string>> list = TCsvFile.GetListOfStringList(path);
+                //List<List<string>> list = FileHelper.CSV.GetListOfStringList(path);
                 //LocalLogger(TLog.LogLevel.DEBUG, "Number of records = {0}", list?.Count);
                 //if ((list?.Count ?? 0) > 0)
                 //{
@@ -1676,10 +1676,10 @@ namespace VsCSharpWinForm_sample2
                 //}
 
                 ///// Example 1.
-                //int recordCount = TCsvFile.ReadFileWithImportedAction(path, CsvHandlingMethod1);
+                //int recordCount = FileHelper.CSV.ReadFileWithImportedAction(path, CsvHandlingMethod1);
                 //LocalLogger(TLog.LogLevel.DEBUG, "Number of records = {0}", recordCount);
                 /// Example 2.
-                string[][] arrayOfStringArray = TCsvFile.ReadFileAndGetArrayOfStringArray(path);
+                string[][] arrayOfStringArray = FileHelper.CSV.ReadFileAndGetArrayOfStringArray(path);
                 LocalLogger(TLog.LogLevel.DEBUG, "Number of records = {0}", arrayOfStringArray?.Length);
                 if ((arrayOfStringArray?.Length ?? 0) > 0)
                 {
@@ -1696,7 +1696,7 @@ namespace VsCSharpWinForm_sample2
                     }
                 }
 
-                //List<Student> students = TCsvFile.Read1<Student>(path);
+                //List<Student> students = FileHelper.CSV.Read1<Student>(path);
 
                 /// Write.
                 path = GeneralT.GetDefaultAbsolutePathIfRelative("testCsvFile1.csv");
@@ -1714,7 +1714,7 @@ namespace VsCSharpWinForm_sample2
                 //        int i = 0;
                 //        while (i < recordCount1)
                 //        {
-                //            if (TCsvFile.WriteStringArrayToStreamWriter(sw, arrayOfStringArray[i], false))
+                //            if (FileHelper.CSV.WriteStringArrayToStreamWriter(sw, arrayOfStringArray[i], false))
                 //            { LocalLogger(TLog.LogLevel.DEBUG, "Succeed to write the record {0}", i); }
                 //            else { LocalLogger(TLog.LogLevel.DEBUG, "Fail to write the record {0}", i); }
                 //            i++;
@@ -1729,7 +1729,7 @@ namespace VsCSharpWinForm_sample2
                     int i = 0;
                     while (i < recordCount2)
                     {
-                        if (TCsvFile.WriteToFile(path, arrayOfStringArray[i], false))
+                        if (FileHelper.CSV.WriteToFile(path, arrayOfStringArray[i], false))
                             LocalLogger(TLog.LogLevel.DEBUG, "Succeed to write the record {0}", i);
                         else LocalLogger(TLog.LogLevel.DEBUG, "Fail to write the record {0}", i);
                         i++;
@@ -2302,6 +2302,52 @@ namespace VsCSharpWinForm_sample2
         }
         #endregion
 
+        private void BtnUri_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                Uri uri = new Uri("sftp://10.15.255.5/Dunhill_uk/FCDB_TestData/SFCI_APAC_P_*.csv");
+                string host = uri?.Host;/// 10.10.255.255
+                string remotePath = uri?.AbsolutePath;/// /abc_folder/sub_folder/abc.zip
+                LocalLogger(TLog.LogLevel.DEBUG, "AbsolutePath = {0}", uri?.AbsolutePath);
+                LocalLogger(TLog.LogLevel.DEBUG, "AbsoluteUri = {0}", uri?.AbsoluteUri);
+                LocalLogger(TLog.LogLevel.DEBUG, "Fragment = {0}", uri?.Fragment);
+                LocalLogger(TLog.LogLevel.DEBUG, "Host = {0}", uri?.Host);
+                LocalLogger(TLog.LogLevel.DEBUG, "HostNameType = {0}", uri?.HostNameType);
+                LocalLogger(TLog.LogLevel.DEBUG, "LocalPath = {0}", uri?.LocalPath);
+                LocalLogger(TLog.LogLevel.DEBUG, "OriginalString = {0}", uri?.OriginalString);
+                LocalLogger(TLog.LogLevel.DEBUG, "PathAndQuery = {0}", uri?.PathAndQuery);
+                LocalLogger(TLog.LogLevel.DEBUG, "Port = {0}", uri?.Port);
+                LocalLogger(TLog.LogLevel.DEBUG, "Query = {0}", uri?.Query);
+                LocalLogger(TLog.LogLevel.DEBUG, "Scheme = {0}", uri?.Scheme);
+                LocalLogger(TLog.LogLevel.DEBUG, "Segments = {0}", uri?.Segments);
+                LocalLogger(TLog.LogLevel.DEBUG, "UserEscaped = {0}", uri?.UserEscaped);
+                LocalLogger(TLog.LogLevel.DEBUG, "UserInfo = {0}", uri?.UserInfo);
+            }
+            catch (Exception ex) { LocalLogger(TLog.LogLevel.ERROR, ex.ToString()); }
+        }
+
+        private void BtnArray1_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                string s = "apple.chan;Apple Chan;23456789";
+                string[] array = s?.Split(';');
+                Student student = new Student()
+                {
+                    UniqueName = array?[0],
+                    DisplayName = array?[1],
+                    Phone = array?[2],
+                    Email = array?[3]
+                };
+                LocalLogger(TLog.LogLevel.DEBUG, "UniqueName = {0}", student.UniqueName);
+                LocalLogger(TLog.LogLevel.DEBUG, "DisplayName = {0}", student.DisplayName);
+                LocalLogger(TLog.LogLevel.DEBUG, "Phone = {0}", student.Phone);
+                LocalLogger(TLog.LogLevel.DEBUG, "Email = {0}", student.Email);
+            }
+            catch (Exception ex) { LocalLogger(TLog.LogLevel.ERROR, ex.ToString()); }
+        }
+
         private void BtnTest1_Click(object sender, EventArgs e)
         {
             try
@@ -2322,6 +2368,17 @@ namespace VsCSharpWinForm_sample2
                 LocalLogger(TLog.LogLevel.DEBUG, "Folder = {0}", s2);
                 string s3 = System.IO.Path.Combine("", s);
                 LocalLogger(TLog.LogLevel.DEBUG, "Filepath = {0}", s3);
+                s = "sftp://10.15.255.5/Dunhill_uk/FCDB_TestData/SFCI_APAC_P_*.csv";
+                LocalLogger(TLog.LogLevel.DEBUG, "Original string = {0}", s);
+                LocalLogger(TLog.LogLevel.DEBUG, "Folder = {0}", System.IO.Path.GetDirectoryName(s));
+                LocalLogger(TLog.LogLevel.DEBUG, "Filename = {0}", System.IO.Path.GetFileName(s));
+                LocalLogger(TLog.LogLevel.DEBUG, "Please go to URI.");
+                s = "2021-05-13T08:37:57.187Z";
+                LocalLogger(TLog.LogLevel.DEBUG, "String = {0}", s);
+                if (DateTime.TryParse(s, out DateTime t)) LocalLogger(TLog.LogLevel.DEBUG, "Time = {0:yyyy-MM-dd HH:mm:ss.fff}. UTC time = {1:yyyy-MM-dd HH:mm:ss.fff}", t, t.ToUniversalTime());
+                else LocalLogger(TLog.LogLevel.DEBUG, "Cannot parse to time.");
+                DateTime? t1 = DateTime.TryParse(s, out t) ? t : (DateTime?)null;
+                LocalLogger(TLog.LogLevel.DEBUG, "Time = {0:yyyy-MM-dd HH:mm:ss.fff}. UTC time = {1:yyyy-MM-dd HH:mm:ss.fff}", t1, t1?.ToUniversalTime());
             }
             catch (Exception ex) { LocalLogger(TLog.LogLevel.ERROR, ex.ToString()); }
         }
