+++ VsCSharpWinForm_sample2/VsCSharpWinForm_sample2/VsCSharpWinForm_sample2.csproj
@@ -69,18 +69,16 @@
     <Compile Include="Helpers\CyptoRijndaelT.cs" />
     <Compile Include="Helpers\DbHelper.cs" />
     <Compile Include="Helpers\ExcelHelper.cs" />
-    <Compile Include="Helpers\FileTransferHelper.cs" />
+    <Compile Include="Helpers\FileHelper.cs" />
     <Compile Include="Helpers\GeneralT.cs" />
     <Compile Include="Helpers\MailHelper.cs" />
     <Compile Include="Helpers\Singleton.cs" />
     <Compile Include="Helpers\SQLiteConfiguration.cs" />
-    <Compile Include="Helpers\TCsvFile.cs" />
     <Compile Include="Helpers\TIniFile.cs" />
     <Compile Include="Helpers\TLog.cs" />
     <Compile Include="Helpers\TTcpClientSocket.cs" />
     <Compile Include="Helpers\TTcpServerSocket.cs" />
     <Compile Include="Helpers\TTcpSocket.cs" />
-    <Compile Include="Helpers\ZipHelper.cs" />
     <Compile Include="Models\MailItem.cs" />
     <Compile Include="Models\MyDbContext.cs" />
     <Compile Include="Models\Param.cs" />
